modules: SX126x: Remove standby error checking
The SX126x status is an unreliable way to detect errors. It is reported
in forums online that it reports errors even when operations are succssful
[1]. The Semtech reference implementation never checks the status value [2].

So let's remove the error checking when setting standby(). In my testing it
incorrectly reports errors. If we want to perform error checking we can use
the GET_DEVICE_ERRORS command or the status after the standby() command instead.

1: https://forum.lora-developers.semtech.com/t/sx1262-failure-to-execute-command/269/11
2: https://github.com/search?q=repo%3ALora-net%2FLoRaMac-node+SX126xGetStatus&type=code

Signed-off-by: Alistair Francis <alistair@alistair23.me>